% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{add_bands}
\alias{add_bands}
\title{Assign a band to the values of a numerical column}
\usage{
add_bands(
  data,
  minutes_column,
  band_cuts = c(0, 1, 5, 10, 25, 50, 100, 150, 200, 300, 500, 750, 1000, 1500, 2000),
  bands_column_name = "minutes_band",
  truncate_in_data_max = FALSE,
  number_bands = FALSE
)
}
\arguments{
\item{data}{Table to add band}

\item{minutes_column}{Name of the numerical column.}

\item{band_cuts}{Numerical vector specifying the cuts of the bands.}

\item{bands_column_name}{Name of the column to be created.}

\item{number_bands}{Whether or not to add numbering to the bands.}
}
\value{
a column with added values
}
\description{
Assign a band to the values of a numerical column
}
\examples{
\dontrun{
add_bands(data, "minutes",
c(0, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60))
}
}
