% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complete_team_names}
\alias{complete_team_names}
\title{Complete names}
\format{
A data frame with 32 rows and 2 variables:
\describe{
  \item{team}{Short Name}
  \item{complete_name}{Complete Name}
  ...
}
}
\usage{
complete_team_names
}
\description{
Complete names
}
\keyword{datasets}
