% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{draw_circle}
\alias{draw_circle}
\title{Dibujar un círculo.
Tomado de: http://stackoverflow.com/questions/6862742/draw-a-circle-with-ggplot2}
\usage{
draw_circle(
  center = c(0, 0),
  diameter = 1,
  npoints = 12000,
  start = 0,
  end = 2
)
}
\arguments{
\item{center}{Centro del círculo}

\item{diameter}{Diámetro del círculo}

\item{npoints}{Cantidad de puntos dibujados}

\item{start}{Inicio del círculo (Va de 0 a 2pi)}

\item{end}{Fin del círculo (Va de 0 a 2pi)}
}
\value{

}
\description{
Dibujar un círculo.
Tomado de: http://stackoverflow.com/questions/6862742/draw-a-circle-with-ggplot2
}
