% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{draw_goal}
\alias{draw_goal}
\title{Draw the goal}
\usage{
draw_goal(color = "#b90d16")
}
\arguments{
\item{color}{Colour of the goal. By default red.}
}
\value{
ggplot plot showing a goal
}
\description{
Draw the goal
}
\examples{
draw_goal("blue")
}
