% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{draw_shots_on_half_court}
\alias{draw_shots_on_half_court}
\title{Takes a data.table with shots, locations (or coordinates) and if goal or not and displays them in the court}
\usage{
draw_shots_on_half_court(
  data,
  ihf_pbp = FALSE,
  goal_column = "goal",
  shot_position_column = "shot_position",
  ...
)
}
\arguments{
\item{data}{Shot data. If ihf_pbp = TRUE, `in_goal_position_column` and `goal_column` so that the funcions fills with the x, y coordinates.}

\item{ihf_pbp}{Should be TRUE if we get the shots data from IHF Play by Play. If not, we need the x and y coordinates columns as `x` and `y`. goal_column (categorical "0" - "1" ) is needed.}

\item{goal_column}{Name of the column that indicates if there is a goal or not (should be in categorical "0" - "1" format)}

\item{shot_position_column}{Name of the column that indicates the in court position (if data has the IHF PBP format)}

\item{...}{Aditional parameters to pass to `handbaloner::half_court()`}
}
\value{
Plot with the goal and the shots displayed (green are goals and red not goals)
}
\description{
Takes a data.table with shots, locations (or coordinates) and if goal or not and displays them in the court
}
\examples{
draw_shots_on_half_court(data, TRUE)
}
