% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{fill_nas}
\alias{fill_nas}
\title{Function used to fill NAs in the possesions sections in pbp}
\usage{
fill_nas(table, column)
}
\arguments{
\item{tabla}{Tabla que recibe}

\item{columna}{Columna a llenar los NAs}
}
\description{
Function used to fill NAs in the possesions sections in pbp
}
\examples{
fill_nas(table, 'numero_jugada')
}
