% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbyplay.R
\name{generar_pbp_tidy}
\alias{generar_pbp_tidy}
\title{Convierte la planilla Play by Play de la página de la IHF en una tabla tidy}
\usage{
generar_pbp_tidy(
  input,
  two_min = "2-minutes suspension",
  match_id_pattern = "Match No: ",
  womens_string_identifier = "Women",
  initial_goalkeepers_together = TRUE
)
}
\arguments{
\item{input}{Directorio donde se encuentra el archivo pbp.}
}
\value{
Una tabla con el partido en formato tidy (una observación por fila)
#@export
}
\description{
Convierte la planilla Play by Play de la página de la IHF en una tabla tidy
}
\examples{
generar_pbp_tidy('01pbp.pdf')
}
