% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbyplay.R
\name{generate_tidy_pbp}
\alias{generate_tidy_pbp}
\title{Converts the play by play sheet from the IFH site to a tidy table}
\usage{
generate_tidy_pbp(
  input,
  two_min = "2-minutes suspension",
  columns_in_spanish = FALSE
)
}
\arguments{
\item{input}{Path of the Play by Play PDF file}

\item{two_min}{How the 2 minutes suspension is described}

\item{columns_in_spanish}{By default, the language is english. If TRUE, the colnames will be in spanish}
}
\value{
Una data.table with the play by play of the game in tidy format (one observation per row)
}
\description{
Converts the play by play sheet from the IFH site to a tidy table
}
\examples{
generate_tidy_pbp('01pbp.pdf')
}
