% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{half_court}
\alias{half_court}
\title{Creación de medio campo verticalmente (arriba)}
\usage{
half_court(
  vertical = TRUE,
  flip = FALSE,
  court_color = "#1871c9",
  area_color = "#d1b111",
  lines_color = "white"
)
}
\arguments{
\item{lines_color}{}
}
\value{

}
\description{
Creación de medio campo verticalmente (arriba)
}
