% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handbaloner-package.R
\docType{package}
\name{handbaloner-package}
\alias{handbaloner}
\alias{handbaloner-package}
\title{handbaloner: Utilidades para el análisis de datos de balonmano}
\description{
Utilidades para el analisis de datos de balonmano. Se generan campos y medios campos con ggplot.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/telaroz/handbaloner}
  \item Report bugs at \url{https://github.com/telaroz/handbaloner/issues}
}

}
\author{
\strong{Maintainer}: Valery Zúñiga Kondrashov \email{telaroz@gmail.com}

}
\keyword{internal}
