% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbyplay.R
\name{jug_y_dt}
\alias{jug_y_dt}
\title{Jugadores y Cuerpo Técnico de un partido en particular}
\usage{
jug_y_dt(input)
}
\arguments{
\item{input}{Directorio donde se encuentra el archivo pbp.}
}
\value{
Una lista de dos tablas: jugadores y cuerpo técnico
}
\description{
Jugadores y Cuerpo Técnico de un partido en particular
}
\examples{
jug_y_dt('01pbp.pdf')
}
