% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineas.R
\name{lines_generator}
\alias{lines_generator}
\title{Genera las líneas necesarias para los plots}
\usage{
lines_generator()
}
\value{
Simplemente genera las líneas necesarias para los gráficos.
}
\description{
Genera las líneas necesarias para los plots
}
