% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playbyplay.R
\name{players_and_coaches}
\alias{players_and_coaches}
\title{Players and coaches}
\usage{
players_and_coaches(input, columns_in_spanish = FALSE)
}
\arguments{
\item{input}{Directory where the pbp is located}
}
\value{
A list with two tables: players and coaches
}
\description{
Players and coaches
}
\examples{
players_and_coaches('01pbp.pdf')
}
