% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_paces}
\alias{plot_paces}
\title{Generate a pace plot for each 5-minutes inverval for 2 teams}
\usage{
plot_paces(pbp_data, chosen_match_id, move_explanation_right = 0)
}
\arguments{
\item{pbp_data}{PBP data}

\item{chosen_match_id}{Which match_id you want to plot}

\item{move_explanation_right}{Explanation on what pace is can be moved to the rigth}
}
\value{
plot with the paces
}
\description{
Generate a pace plot for each 5-minutes inverval for 2 teams
}
\examples{
plot_paces(pbp_data, 1)
}
