% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{scrape_from_ihf}
\alias{scrape_from_ihf}
\title{Function to download pdfs of World Championships in the new IHF site}
\usage{
scrape_from_ihf(link, folder, from_archive = FALSE)
}
\arguments{
\item{link}{Site were we want to download the pdfs}

\item{folder}{Folder were the pdfs will be downloaded}

\item{from_archive}{Whether the pdfs are in the archive.ihf.info site}
}
\description{
Function to download pdfs of World Championships in the new IHF site
}
\examples{
scrape_from_ihf(link = 'https://www.ihf.info/competitions/men/308/27th-ihf-men039s-world-championship-2021/22415/match-center/23765', carpeta = 'matches')
}
