% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xg_egipto21}
\alias{xg_egipto21}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{posicion_tiro}{Position of the shot}
  \item{xg}{Expected Goals}
  \item{cantidad_tiros}{Number of shot in this position in this zone}
  ...
}
}
\source{
\url{'https://raw.githubusercontent.com/telaroz/egipto21/main/partidos_pbp_egipto2021.csv'}
}
\usage{
xg_egipto21
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
