% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{court}
\alias{court}
\title{Creación del campo completo}
\usage{
court(
  vertical = FALSE,
  flip = FALSE,
  court_color = "#1871c9",
  area_color = "#d1b111",
  lines_color = "white"
)
}
\value{

}
\description{
Creación del campo completo
}
