% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{distance_to_goal}
\alias{distance_to_goal}
\title{Calculate the distance to goal in any point of a handbaloner::court()}
\usage{
distance_to_goal(shot_coordinates)
}
\arguments{
\item{shot_coordinates}{}
}
\value{

}
\description{
Calculate the distance to goal in any point of a handbaloner::court()
}
