% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{distance_to_goal}
\alias{distance_to_goal}
\title{Calcula la distancia a gol desde unas coordenadas en handbaloner::court()}
\usage{
distance_to_goal(x, y)
}
\arguments{
\item{x}{Coordenada x (centro del campo en (0, 0))}

\item{y}{Coordenada y (centro del campo en (0, 0))}
}
\value{
double
}
\description{
Toma el centro del campo como las coordenadas (0, 0) y calcula la distancia
a gol desde un punto en específico. El cálculo se toma desde el punto al gol
más cercano, es decir, desde el propio campo. Un tiro del campo contrario no
lo calculará correctamente.
}
\examples{
\dontrun{
distance_to_goal(x = -10, y = -3)
}
}
