% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xg_egipto21}
\alias{xg_egipto21}
\title{Expected goals from the complete egypt 2021 men's world championship}
\format{
A data frame with 12 rows and 3 variables:
\describe{
  \item{posicion_tiro}{Position of the shot}
  \item{xg}{Expected Goals}
  \item{cantidad_tiros}{Number of shot in this position in this zone}
  ...
}
}
\source{
\url{'https://raw.githubusercontent.com/telaroz/egipto21/main/partidos_pbp_egipto2021.csv'}
}
\usage{
xg_egipto21
}
\description{
Expected goals from the complete egypt 2021 men's world championship
}
\keyword{datasets}
